requirements.txt

---

## ✅ `requirements.txt`

```txt
pandas
matplotlib
scikit-learn
streamlit
python-dateutil
